Move example fps display to test mod. There may be a discussion about this, but my argument for this is that a lib shouldn't also implement an example, only in the test mod. IMO this should be moved to the generic hud mod.
