+++ src/testmod/java/dev/wooferz/hudlib/test/BaseTextElement.java
@@ -1,4 +1,4 @@
-package dev.wooferz.hudlib.test.element;
+package dev.wooferz.hudlib.test;
 
 import dev.isxander.yacl3.api.Option;
 import dev.isxander.yacl3.api.OptionGroup;
@@ -7,8 +7,6 @@ import dev.isxander.yacl3.api.controller.TickBoxControllerBuilder;
 import dev.wooferz.hudlib.HudAnchor;
 import dev.wooferz.hudlib.hud.HUDConfig;
 import dev.wooferz.hudlib.hud.HUDElement;
-import dev.wooferz.hudlib.test.BaseTextConfig;
-import dev.wooferz.hudlib.test.HudLibTest;
 import dev.wooferz.hudlib.utils.TextUtils;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.font.TextRenderer;
