+++ src/testmod/java/dev/wooferz/hudlib/test/CoordinateElement.java
@@ -1,4 +1,4 @@
-package dev.wooferz.hudlib.test.element;
+package dev.wooferz.hudlib.test;
 
 import dev.wooferz.hudlib.HudAnchor;
 import net.minecraft.client.MinecraftClient;
