+++ src/testmod/java/dev/wooferz/hudlib/test/HudLibTest.java
@@ -1,7 +1,6 @@
 package dev.wooferz.hudlib.test;
 
 import dev.wooferz.hudlib.HudManager;
-import dev.wooferz.hudlib.test.element.CoordinateElement;
 import net.fabricmc.api.ClientModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,5 +13,7 @@ public class HudLibTest implements ClientModInitializer {
     public void onInitializeClient() {
         HudManager.registerHudElement(new CoordinateElement());
 
+        ExampleHUDElement exampleHUDElement = new ExampleHUDElement(5, 5, 55, 17, 1);
+        HudManager.registerHudElement(exampleHUDElement);
     }
 }
